requirements.txt
opencv-python
tqdm
h5py
pandas
scikit-image
git+https://github.com/ercanburak/PerceptualSimilarity
einops
GitPython
wandb
ffmpeg-python
yachalk
moviepy
imageio
git+https://github.com/rtqichen/torchdiffeq
graphviz
torchviz
openpyxl
pytorch-msssim
thop
piq
opencv-contrib-python
git+https://github.com/chaofengc/IQA-PyTorch.git@08f88508f71eaf02234ebd04aeda091fdc3ab8e7